unify generate scripts: called from central generate.py
- renamed old generate.py to generate_from_statistics
- adjusted generate test to use new generate_from_statistics
- fixed example generate_from_csv (vehicle type was missing)
