+++ examples/configs/generate_from_csv.cfg
@@ -2,6 +2,8 @@
 # call generate script with --config option to use this file:
 # python generate.py --config examples/generate.cfg
 
+mode = csv
+
 # set output file
 output = examples/scenario.json
