+++ examples/configs/generate_from_simbev.cfg
@@ -2,6 +2,8 @@
 # call generate_from_simbev script with --config option to use this file:
 # python generate_from_simbev.py --config examples/generate_from_simbev.cfg
 
+mode = simbev
+
 # set output file
 output = simbev_scenario.json
