+++ examples/configs/generate_from_statistics.cfg
@@ -2,6 +2,8 @@
 # call generate script with --config option to use this file:
 # python generate.py --config examples/generate.cfg
 
+mode = statistics
+
 # set output file
 output = examples/scenario.json
 # set vehicle types file
